fix: clear the CI gates the artifacts diff tripped
turn_index becomes bigint (squawk's int rule); the redundant test cast
drops the type-suppression count back to its floor; the composite-FK
catalog test counts artifact as the 19th table; and the changed code's
tests now kill mutants past both patch-mutation thresholds — the PR-files
provider read gets a direct contract suite, readPrArtifacts gets exact
scoping/ordering/fallback proofs, the artifact reconciler gets activity-
window boundaries, recycled-branch tiebreaks, the single-candidate
fallback and grace age-out, the renderer gets overflow caps, label
escaping, proof joins and criterion ordering, refresh proves the
artifact-only presence path, and the gateway route pins its OpenAPI
contract, the request-size refusal, branch-only anchoring, and the
no-PR no-nomination rule.

+++ apps/tenant-dashboard/supabase/migrations/20260814161703_add_artifact.sql
@@ -25,7 +25,7 @@
     "provenance" text not null,
     "session_id" text not null default ''::text,
     "trace_id" text not null default ''::text,
-    "turn_index" integer,
+    "turn_index" bigint,
     "repository" text not null default ''::text,
     "pr_number" bigint,
     "git_repo" text not null default ''::text,
